reorganize in preparation for nav bar changes
