+++ lib/navigation/app_router.dart
@@ -2,8 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:jellyflix/models/auth_state.dart';
-
-import 'package:jellyflix/components/responsive_navigation_bar.dart';
+import 'package:jellyflix/components/jfx_nav_bar/jfx_nav_bar.dart';
 import 'package:jellyflix/models/screen_paths.dart';
 import 'package:jellyflix/providers/auth_provider.dart';
 import 'package:jellyflix/providers/connectivity_provider.dart';
@@ -40,7 +39,7 @@ class AppRouter {
           pageBuilder: (context, state, child) {
             //print(state.location);
             return NoTransitionPage(
-                child: ResponsiveNavigationBar(
+                child: JfxNavBar(
               body: child,
             ));
           },
